DVS使用時のクロスフェーダー遅延補正を追加

+++ STM32CubeIDE/JUMBLEQ/Appli/Core/Inc/SigmaStudioFW.h
@@ -35,6 +35,7 @@ typedef unsigned char ADI_REG_TYPE;
 #define SIGMASTUDIOTYPE_INTEGER  1
 
 #define DEVICE_ADDR_ADAU146XSCHEMATIC_1 0x00
+// Larger DSP memory downloads are split into word-aligned blocks internally.
 #define SIGMA_WRITE_BLOCK_MAX_PAYLOAD    4093U
 
 /*
@@ -49,7 +50,7 @@ typedef unsigned char ADI_REG_TYPE;
  * Write to multiple Device registers
  */
 // #define SIGMA_WRITE_REGISTER_BLOCK( devAddress, address, length, pData ) {/*TODO: implement macro or define as function*/}
-void SIGMA_WRITE_REGISTER_BLOCK(uint8_t devAddress, uint16_t address, uint16_t length, uint8_t* pData);
+void SIGMA_WRITE_REGISTER_BLOCK(uint8_t devAddress, uint16_t address, uint32_t length, uint8_t* pData);
 void SIGMA_WRITE_REGISTER_BLOCK_IT(uint8_t devAddress, uint16_t address, uint16_t length, uint8_t* pData);
 
 // Diagnostics for runtime DSP parameter write path (IT mode)
